backend/requirements/requirements.txt
Django==1.9.5
psycopg2==2.6.1
djangorestframework==3.3.3
